+++ radix-engine/src/transaction/builder.rs
@@ -355,7 +355,7 @@ impl<'a, A: AbiProvider> TransactionBuilder<'a, A> {
         .0
     }
 
-    fn single_authority(badge: Address, permission: u16) -> HashMap<Address, u16> {
+    fn single_authority(badge: Address, permission: u64) -> HashMap<Address, u64> {
         let mut map = HashMap::new();
         map.insert(badge, permission);
         map
@@ -375,7 +375,7 @@ impl<'a, A: AbiProvider> TransactionBuilder<'a, A> {
                 scrypto_encode(&ResourceType::Fungible { divisibility: 18 }),
                 scrypto_encode(&metadata),
                 scrypto_encode(&(MINTABLE | BURNABLE)),
-                scrypto_encode(&0u16),
+                scrypto_encode(&0u64),
                 scrypto_encode(&Self::single_authority(
                     mint_badge_address,
                     MAY_MINT | MAY_BURN,
@@ -399,9 +399,9 @@ impl<'a, A: AbiProvider> TransactionBuilder<'a, A> {
             args: vec![
                 scrypto_encode(&ResourceType::Fungible { divisibility: 18 }),
                 scrypto_encode(&metadata),
-                scrypto_encode(&0u16),
-                scrypto_encode(&0u16),
-                scrypto_encode(&HashMap::<Address, u16>::new()),
+                scrypto_encode(&0u64),
+                scrypto_encode(&0u64),
+                scrypto_encode(&HashMap::<Address, u64>::new()),
                 scrypto_encode(&Some(NewSupply::Fungible {
                     amount: initial_supply.into(),
                 })),
@@ -424,7 +424,7 @@ impl<'a, A: AbiProvider> TransactionBuilder<'a, A> {
                 scrypto_encode(&ResourceType::Fungible { divisibility: 0 }),
                 scrypto_encode(&metadata),
                 scrypto_encode(&(MINTABLE | BURNABLE)),
-                scrypto_encode(&0u16),
+                scrypto_encode(&0u64),
                 scrypto_encode(&Self::single_authority(
                     mint_badge_address,
                     MAY_MINT | MAY_BURN,
@@ -448,9 +448,9 @@ impl<'a, A: AbiProvider> TransactionBuilder<'a, A> {
             args: vec![
                 scrypto_encode(&ResourceType::Fungible { divisibility: 0 }),
                 scrypto_encode(&metadata),
-                scrypto_encode(&0u16),
-                scrypto_encode(&0u16),
-                scrypto_encode(&HashMap::<Address, u16>::new()),
+                scrypto_encode(&0u64),
+                scrypto_encode(&0u64),
+                scrypto_encode(&HashMap::<Address, u64>::new()),
                 scrypto_encode(&Some(NewSupply::Fungible {
                     amount: initial_supply.into(),
                 })),
