+++ scrypto/src/resource/resource_builder.rs
@@ -14,9 +14,9 @@ pub const DIVISIBILITY_MAXIMUM: u8 = 18;
 pub struct ResourceBuilder {
     resource_type: ResourceType,
     metadata: HashMap<String, String>,
-    flags: u16,
-    mutable_flags: u16,
-    authorities: HashMap<Address, u16>,
+    flags: u64,
+    mutable_flags: u64,
+    authorities: HashMap<Address, u64>,
 }
 
 impl ResourceBuilder {
@@ -54,19 +54,19 @@ impl ResourceBuilder {
     }
 
     /// Sets the feature flags.
-    pub fn flags(&mut self, flags: u16) -> &mut Self {
+    pub fn flags(&mut self, flags: u64) -> &mut Self {
         self.flags = flags;
         self
     }
 
     /// Sets the features flags that can be updated in future.
-    pub fn mutable_flags(&mut self, mutable_flags: u16) -> &mut Self {
+    pub fn mutable_flags(&mut self, mutable_flags: u64) -> &mut Self {
         self.mutable_flags = mutable_flags;
         self
     }
 
     /// Adds a badge for authorization.
-    pub fn badge<A: Into<ResourceDef>>(&mut self, badge_address: A, permissions: u16) -> &mut Self {
+    pub fn badge<A: Into<ResourceDef>>(&mut self, badge_address: A, permissions: u64) -> &mut Self {
         self.authorities
             .insert(badge_address.into().address(), permissions);
         self
